less casting

+++ testlib/basic.c
@@ -3,7 +3,7 @@
 extern EXPORT int
 return_passed_integer_value(int input)
 {
-  return (int) input;
+  return input;
 }
 
 extern EXPORT int
@@ -21,23 +21,23 @@ sum_integer_values10(int a1, int a2, int a3, int a4, int a5, int a6, int a7, int
 extern EXPORT long
 char_to_long(signed char input)
 {
-  return (long) input;
+  return input;
 }
 
 extern EXPORT long
 short_to_long(short input)
 {
-  return (long) input;
+  return input;
 }
 
 extern EXPORT long
 int_to_long(int input)
 {
-  return (long) input;
+  return input;
 }
 
 extern EXPORT long
 long_to_long(long input)
 {
-  return (long) input;
+  return input;
 }
